+++ backend/lighthouse/src/main/java/com/ssafy/lighthouse/domain/study/repository/StudyNoticeRepository.java
@@ -10,5 +10,5 @@ import com.ssafy.lighthouse.domain.study.entity.StudyNotice;
 
 public interface StudyNoticeRepository extends JpaRepository<StudyNotice, Integer> {
 	@Query(value = "SELECT sn FROM StudyNotice sn WHERE  sn.isValid = 1 AND sn.studyId = :studyId")
-	public List<StudyNotice> findByStudyId(@Param("studyId") int studyId);
+	public List<StudyNotice> findByStudyId(@Param("studyId") Long studyId);
 }
