+++ backend/lighthouse/src/main/java/com/ssafy/lighthouse/domain/study/service/StudyNoticeService.java
@@ -6,10 +6,10 @@ import com.ssafy.lighthouse.domain.study.dto.StudyNoticeDto;
 import com.ssafy.lighthouse.domain.study.entity.StudyNotice;
 
 public interface StudyNoticeService {
-	List<StudyNotice> findAllByStudyId(int studyId);
-	int createNotice(StudyNoticeDto.StudyNoticeReq dto);
-	int updateNotice(int id, StudyNoticeDto.StudyNoticeReq dto);
-	int removeNotice(int id);
-	int createNoticeCheck(StudyNoticeDto.StudyNoticeCheckReq dto);
-	StudyNotice findById(int id);
+	List<StudyNotice> findAllByStudyId(Long studyId);
+	Long createNotice(StudyNoticeDto.StudyNoticeReq dto);
+	Long updateNotice(Long id, StudyNoticeDto.StudyNoticeReq dto);
+	Long removeNotice(Long id);
+	Long createNoticeCheck(StudyNoticeDto.StudyNoticeCheckReq dto);
+	StudyNotice findById(Long id);
 }
